+++ scene/src/com/hendraanggrian/openpss/control/StretchableSplitMenuButton.kt
@@ -3,8 +3,8 @@
 package com.hendraanggrian.openpss.control
 
 import javafx.beans.DefaultProperty
-import javafx.beans.property.IntegerProperty
-import javafx.beans.property.SimpleIntegerProperty
+import javafx.beans.property.DoubleProperty
+import javafx.beans.property.SimpleDoubleProperty
 import javafx.beans.property.SimpleStringProperty
 import javafx.beans.property.StringProperty
 import javafx.scene.Node
@@ -12,13 +12,13 @@ import javafx.scene.control.SplitMenuButton
 
 @DefaultProperty("graphic")
 class StretchableSplitMenuButton @JvmOverloads constructor(
-    stretchPoint: Int = -1,
+    stretchPoint: Double = -1.0,
     stretchableText: String? = null,
     graphic: Node? = null
 ) : SplitMenuButton(), StretchableLabeled {
 
-    private val stretchPointProperty = SimpleIntegerProperty(stretchPoint)
-    override fun stretchPointProperty(): IntegerProperty = stretchPointProperty
+    private val stretchPointProperty = SimpleDoubleProperty(stretchPoint)
+    override fun stretchPointProperty(): DoubleProperty = stretchPointProperty
 
     private val stretchableTextProperty = SimpleStringProperty(stretchableText)
     override fun stretchableTextProperty(): StringProperty = stretchableTextProperty
